Move AddEntryToTree to UrlMatchingTree
